split hnsw and hnsw batch into declaration and imp files, removed new batch iterator function for the factory
